Removed Timed Out messages, added thread joining logic to prevent pre…

+++ src/dibas_fits_writer/BfFitsIO.h
@@ -116,11 +116,11 @@ public:
     virtual void createPrimaryHDU();
     void createDataTable();
 
-    int writeRow(int mcnt,int good_data, float *data, bool cmp);
-    int write_HI(int mcnt,int good_data, float *data);
-    int write_PAF(int mcnt,int good_data, float *data);
-    int write_FRB(int mcnt,int good_data, float *data);
-    int write_RTBF(int mcnt,int good_data, float *data);
+    int writeRow(  int mcnt, int64_t good_data, float *data, bool cmp);
+    int write_HI(  int mcnt, int64_t good_data, float *data);
+    int write_PAF( int mcnt, int64_t good_data, float *data);
+    int write_FRB( int mcnt, int64_t good_data, float *data);
+    int write_RTBF(int mcnt, int64_t good_data, float *data);
 
     bool is_scan_complete(int mcnt);
     void set_scan_complete();
