+++ src/dibas_fits_writer/BfFitsThread.cc
@@ -317,14 +317,14 @@ BfFitsThread::run(struct vegas_thread_args *args)
 
         // collect some mode dependent info about the databuffer blocks
         uint64_t mcnt,n_block;
-        int gd;
+        int64_t gd;
         float *data;
         if (cov_mode1) {
             mcnt = ((bf_databuf *)gdb)->block[block].header.mcnt;
             n_block = ((bf_databuf *)gdb)->header.n_block;
             gd = ((bf_databuf *)gdb)->block[block].header.good_data;
             data = ((bf_databuf *)gdb)->block[block].data;
-            fitsio->write_HI(mcnt,gd, data);
+            fitsio->write_HI(mcnt, gd, data);
 	    printf("mcnt: %llu\n",(long long unsigned int) mcnt);
         }
 
@@ -333,7 +333,7 @@ BfFitsThread::run(struct vegas_thread_args *args)
             n_block = ((bfpaf_databuf *)gdb)->header.n_block;
             gd = ((bf_databuf *)gdb)->block[block].header.good_data;
             data = ((bfpaf_databuf *)gdb)->block[block].data;
-            fitsio->write_PAF(mcnt, gd,data);
+            fitsio->write_PAF(mcnt, gd, data);
 	    printf("mcnt: %llu\n",(long long unsigned int) mcnt);
         }
 
@@ -342,7 +342,7 @@ BfFitsThread::run(struct vegas_thread_args *args)
             n_block = ((bffrb_databuf *)gdb)->header.n_block;
             gd = ((bf_databuf *)gdb)->block[block].header.good_data;
             data = ((bffrb_databuf *)gdb)->block[block].data;
-            fitsio->write_FRB(mcnt,gd, data);
+            fitsio->write_FRB(mcnt, gd, data);
             printf("mcnt: %llu\n",(long long unsigned int) mcnt);
         }
         else {
@@ -352,7 +352,7 @@ BfFitsThread::run(struct vegas_thread_args *args)
             data = ((bfp_databuf *)gdb)->block[block].data;
             printf("mcnt: %llu\n",(long long unsigned int) mcnt);
             num_iter++;
-            fitsio->write_RTBF(mcnt,gd, data);
+            fitsio->write_RTBF(mcnt, gd, data);
         }    
         clock_gettime(CLOCK_MONOTONIC, &fits_stop);
         total_write_time += ELAPSED_NS(fits_start, fits_stop);
